code/requirements.txt
# AI OS Personal - Content Quality Loop
# Python 3.9+

# Gemini API
google-genai==1.61.0

# Environment variables
python-dotenv==1.0.0

# Task queue (Background Mode)
celery==5.3.4
redis==5.0.1

# CLI interface
click==8.1.7

# Email notifications
python-smtp==1.0.0

# Date/time utils
python-dateutil==2.8.2

# JSON processing
jsonschema==4.20.0

# Optional: Notion integration
# notion-client==2.2.1

# Optional: Slack notifications
# slack-sdk==3.26.1
